Add `type` attribute to A_CONST nodes.
This fixes an issue where BitString's cannot be deparsed properly
because they cannot be distinguished from normal strings.

Example:

SELECT B'0101'

and

SELECT '0101'

This patch adds a `type` which outputs the integer node tag so that
callers can be certain of the type of the constant given the limitations
of the serialization format (JSON). This also fixes a bug where
`BitString`s were being dumped as invalid JSON:

```
{ "val": b0101 }
```