Removed references to 128 bit
This makes it compatible with 32 bit systems. We don't need use support
for 128 bit integers anywhere in the codebase.

An alternative option would be to make this depend on a configure script
(like postgres does to generate pg_config.h), but this seems overkill
given we don't really have need for them.