Normalize: Adjust logic for DefElem to work without parser patch
Previously we relied on a parser modification that would create A_Const
for DefElems in order to locate constant values for DefElem which have
a String argument.

In general, its good to reduce the number of parser patches, and this
particular case can be directly handled in the normalize method instead.

Note that the most prominent reason for doing this are password-related
utility commands, such as "ALTER USER test PASSWORD 'xyz'", which is
unfortunately not parameterized by pg_stat_statement's own logic.