Add experimental Protobuf C++ outfuncs, converge JSON output to match
This introduces the following changes to the JSON output format, in order
to match the Protobuf JSON mapping scheme:

- Default values are no longer output for Enum values
- Whitespace between control characters in JSON is no longer added
- "<" and ">" characters are escaped to avoid browser HTML injections

Separately, this commit also unifies all outfuncs related code into the
"pg_query_outfuncs" prefixed files, to make it easier to understand.

The Protobuf C++ build can be utilized by building with
"make USE_PROTOBUF_CPP=1". Note that the Protobuf C++ requirement is for
building only, and the library itself remains a C library (with a bundled
version of Protobuf-C).