Free Postgres top-level memory context on thread exit / with function
Previously there was no way to free the top-level Postgres memory context,
causing threaded programs that churn through a lot of threads to leak
memory with each newly initialized thread-local top-level memory context.

Instead, this uses a newly introduced cleanup method to free the memory
when a pthread exits (note this causes a pthread dependency to be added
to this library). In addition, primarily for memory testing purposes, add a
new method "pg_query_exit" that performs the same cleanup on demand.