JSON outfuncs: Fix handling of NIL list elements
Certain parse nodes intentionally fill a list with a NIL element, and
we didn't consistently output this the same way across Protobuf and
JSON implementations.

This is now output as an empty node, which should be considered a null
value. In protobuf this is represented as a Node message without a set
type (PG_QUERY__NODE__NODE__NOT_SET for Protobuf-C).