Improve deparser based on Postgres regression test files
This is a start to ensure our deparser works correctly by making use
of the SQL files provided in the Postgres source tree as part of the
regression test suite.

Note that for the regression tests we do not compare the output string,
because its not possible to return the exact same statement text due
to details such as whitespace and comments. However, we can compare
that parsing the resulting, deparsed query again yields an identical
parse tree (minus the location fields).