Use XXH-based 64-bit integer for fingerprint, instead of SHA1-based string
This enables faster computation / better storage of the fingerprint, and
doesn't actually affect its usefulness much, because the typical
underlying data source for fingerprinted data (pg_stat_statements) also
uses a 64-bit integer to identify queries.

In addition, this vendors the xxhash library for computing the fingerprint,
and stops using the cryptographic SHA1 hash function for this, which is
unnecessarily slow. Note we don't use the Postgres hash function here,
because unlike pg_stat_statements we need the fingerprint to be portable
across different architectures.

For future proofing, all vendored libraries (protobuf-c and xxhash) are
now located in the new vendor/ directory. When not using the created
static library, make sure to include both protobuf-c and xxhash in the
build process for any libpg_query-based libraries.