Fingerprint: Ignore TypeCast when the arg is A_Const or ParamRef
This ensures a few edge cases are handled correctly, specifically
"SELECT a = $1" should have the same fingerprint as "SELECT a = false".

Whilst there might be value in considering typecasts as something that
is impactful enough to be included in the fingerprint, for now it seems
better to err on the side of too much grouping, and ignore all typecasts
if they contain a constant value or parameter reference.