Fingerprinting: Ignore multiple subsequent digits in RangeVar.relname
This ensures that time-based partitioned tables, which usually include
the date in the table partition name, are treated as the same.

To avoid grouping together things like "table_v1" and "table_v2", we
require at least two subsequent digits.

This logic isn't perfect, but seems to cover the most common cases.