Fingerprinting: Optimize list and node field fingerprinting
For list fingerprints we were comparing the full token list of each node
alphabetically, which can be very expensive. Instead, run two passes of
fingerprinting, once to calculate the hash value into a sub context, and
then sort the list using that, running the fingerprint again with the
parent hash.

In addition, also update the rather inefficient "don't output node field
name if its value is empty" to use a simple backtracking mechanism that
stores the previous hash value.

These two optimizations combined bring the runtime of test/complex on a
2017 iMac Pro from ~7 seconds to ~4 seconds, likely due to the avoidance
of unnecessary string allocations and associated copying.