Fingerprint: Add explicit fingerprint_str result to avoid confusion
In case fingerprints are output, its preferable to output the canonical
form of the hash (XXH64_canonicalFromHash) represented as hex values.

This happens to be identical to printf("%016x", hash), but it seems less
error-prone to go directly through xxHash's canonical methods to protect
against any oversights in regards to endianness.