Fingerprint: Don't output node name when node type is specified in parent
This matches the JSON format, which no longer presents the node names
when the parent node has a fixed type field, instead of the generic Node
type. This ultimately comes from the Protobuf format, which avoids an
unnecessary Node message when a field always has the exact same type
present for a field (the same way this is done in Postgres itself).