Deparser: Fix crash in CopyStmt with HEADER or FREEZE inside WITH parens
The parse tree does not contain an explicit argument in those cases,
but does when specified in the legacy mode without the wrapping WITH.

With this change we only output the "1" argument when the original tree
also had this, to ensure parse tree comparisons match. Note the intent
here is technically the same, which is to enable these options.