Fingerprint: Cache list item hashes to fingerprint complex queries faster
This was exhibiting quite bad runtime behaviour before, causing both an
explosion in memory, as well as very high CPU runtime for complex queries.

The new approach ensures we don't calculate the hashes for a particular
list more than once, which ensures that we roughly have quadratic runtime
instead of exponential runtime.