Normalize: Match GROUP BY against target list and re-use param refs
This avoids "columns ... must appear in the GROUP BY clause ..." type
errors when the normalize result is utilized as a parameterized query
with Postgres. We reuse the pg_query fingerprinting logic to determine
whether a target list element is identical with a GROUP BY element.

Note that query data coming from pg_stat_statements normalizes the target
list, but not the GROUP BY list.

Additionally, this overly optimizes for reusing param refs, specifically
when the target list contains elements that are similar, but not 100%
identical, we err on matching them (based on the pg_query fingerprint),
instead of requiring full equivalency. e.g. the target list may contain
a certain string constant and we would consider any other string constant
in the GROUP BY to be equivalent (since they have the same fingerprint).