Ensure we don't rely on x86 and ARM intrinsics (i.e. SSE 4.2, etc)
Add three missing defines to always be unset to avoid this dependency:
- HAVE_X86_64_POPCNTQ
- USE_ARMV8_CRC32C
- USE_SSE42_CRC32C_WITH_RUNTIME_CHECK

Additionally, to help with generating the correct code on either ARM or
x86, make sure we unset these defines before we run the extract source
script. This now produces the identical output on x86 and ARM, besides
some minor pg_config differences before the undefs at the end that undo
those differences.

In passing, drop the "-msse4.2" flag passed in during source extraction,
as we do not want to target SSE 4.2.