Fix Value node fingerprinting
This commit adjusts fingerprint strings to properly output the
corresponding `sval`, `fval`, and `bsval` field names when
fingerprinting corresponding value nodes. Additionally, `Boolean`
nodes are now output in the fingerprint logic.

The order of some list items was shuffled in the output; this is because
the updated fingerprint strings result in differing hashes for list
items, which are used to sort list items. Fingerprint hashes and expected
tokens were updated accordingly.