Deparser: Add support for COALESCE and other expressions in LIMIT clause
We've previously supported "c_expr" being used in a LIMIT clause, but
its actually more correct to treat this as "a_expr", which is a broader
set of functions, including COALESCE.

Note that the previous handling (for "c_expr") is still correct for
FETCH FIRST (expr) ROWS WITH TIES, see the "select_fetch_first_value"
in the Postgres grammar.

In passing fix two new warnings in the deparser caused by a recent code
addition.