Improve handling of Postgres version-related defines
Previously there used to be a recurring bug where the defines suddenly
became empty when regenerating sources, due to an ordering issue. Fix by
always using the versions defined on top of the Makefile as the source.

Additionally, rewrite PG_VERSION_STR to avoid the string containing
information from the environment the source was extracted in (which
caused unnecessary diffs between different machines).