extract_source: Handle additional includes that are always needed directly
This moves the decision which includes we copy, on top of those we
discover to be needed, from the Makefile to the extract_source script,
to make it easier to manage the list.

In passing this also broadens the patterns of which atomics includes
are kept (previously we only kept x86/ARM/PPC, but it doesn't hurt
to keep all of them).