Windows support: Simplify write_stderr and should_output_to_client
In the pg_query context these do not need to do the work they usually do
(we always output to stderr, and we never output to the client), and
mocking them avoids pulling in postmaster.c amongst other benefits. It
also avoids problems with Windows-specific code in a follow-up commit.