JSON outfuncs: Do not output "ival" key (and similar) twice for A_Const
The JSON format intends to follow the structure of the Protobuf JSON
representation of the parse nodes. The A_Const parse node is the only
one making use of the "oneof" construct (matching the union in the C code)
and was using a different representation in the generated JSON protobuf
(outputting "ival" once), vs the manual JSON output (outputting it twice).

Match the manual JSON output to the generate JSON protobuf. This changes
an output of "{"A_Const":{"ival":{"ival":123}}}" to be
"{"A_Const":{"ival":123}}" instead.

This is a breaking change in the JSON output.