Fix parsing of negative integer constants
Currently the value of negative constants are omitted. Only omit *0*
constants.

An earlier attempt (#244) caused a mismatch with the protobuf json
output, since that wants to omit values that are equal to the default
value.

I don't *really* think that is particularly sensible behavior for an
integer constant in an AST, but oh well.

Fixes #243.