PL/pgSQL: Detect fully qualified argument types outside of pg_catalog as records
This ensures the PL/pgSQL lexer detects variable references to such
variables as variables, instead of erroring out with
"[var.name] is not a known variable".

The trade-off here is that we might incorrectly detect custom data types
registered outside of pg_catalog as being records, instead of data types.